- added delay after each char is transmitted in morse prototype - commented code in morse prototype - created binary prototype in python - worked on c prototype, not much progress
